build: Only remove temporary bundle directories
When the handle_bundle_common() function was introduced, the build
command unintentionally started deleting bundle directories after
combining container images, regardless if the directory was a temporary
one or not. This commit restores the original intended behavior of only
removing bundle directories created when a docker compose yaml file is
passed to the 'bundle' section instead of a complete directory.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>